[RFR] add config for prod

+++ Makefile
@@ -6,6 +6,7 @@ DOCKER_COMPOSE_BUILD = docker-compose -p reaction-rapide-build -f docker-compose
 DOCKER_COMPOSE_TEST = docker-compose -p reaction-rapide-test -f docker-compose.yml -f docker-compose.test.yml
 DOCKER_COMPOSE_E2E = docker-compose -p reaction-rapide-e2e -f docker-compose.yml -f docker-compose.e2e.yml
 DOCKER_COMPOSE_STAGING = docker-compose -p reaction-rapide-staging -f docker-compose.yml -f docker-compose.staging.yml
+DOCKER_COMPOSE_PROD = docker-compose -p reaction-rapide-prod -f docker-compose.yml -f docker-compose.prod.yml
 DOCKER_COMPOSE_DEV_NGINX = docker-compose -p reaction-rapide-dev-nginx -f docker-compose.yml -f docker-compose.dev-nginx.yml
 
 install: install-admin
@@ -20,14 +21,20 @@ install-staging:
 start:
 	$(DOCKER_COMPOSE) up --force-recreate -d
 
+stop:
+	$(DOCKER_COMPOSE) down
+
 start-staging:
 	$(DOCKER_COMPOSE_STAGING) up -d
 
 stop-staging:
 	$(DOCKER_COMPOSE_STAGING) down
 
-stop:
-	$(DOCKER_COMPOSE) down
+start-prod:
+	$(DOCKER_COMPOSE_PROD) up -d
+
+stop-prod:
+	$(DOCKER_COMPOSE_PROD) down
 
 logs:
 	$(DOCKER_COMPOSE) logs -f
@@ -70,6 +77,11 @@ DB_MIGRATE_STAGING = $(DOCKER_COMPOSE_STAGING) run --rm api sh -c "/app/var/wait
 	--migrations-dir=migrations \
 	-e api
 
+DB_MIGRATE_PROD = $(DOCKER_COMPOSE_PROD) run --rm api sh -c "/app/var/wait-for-it.sh -h db -p 5432 -t 30 && ./node_modules/.bin/db-migrate \
+	--config=database.js \
+	--migrations-dir=migrations \
+	-e api
+
 DB_MIGRATE_E2E = $(DOCKER_COMPOSE_E2E) run --rm api sh -c "/app/var/wait-for-it.sh -h db-e2e -p 5432 -t 30 && ./node_modules/.bin/db-migrate \
 	--config=database.js \
 	--migrations-dir=migrations \
@@ -91,6 +103,9 @@ migration-test:
 
 migration-staging:
 	$(DB_MIGRATE_STAGING) up"
+
+migration-prod:
+	$(DB_MIGRATE_PROD) up"
 
 migration-e2e:
 	$(DB_MIGRATE_E2E) up"
@@ -114,7 +129,10 @@ debug-e2e:
 	$(DOCKER_COMPOSE_E2E) run test-e2e
 
 deploy-staging:
-	npx shipit staging deploy
+	NODE_ENV=staging npx shipit staging deploy
+
+deploy-prod:
+	NODE_ENV=production npx shipit production deploy
 
 install-admin:
 	$(DOCKER_COMPOSE) run --rm --no-deps admin npm install
@@ -122,20 +140,30 @@ install-admin:
 build-storybook:
 	$(DOCKER_COMPOSE_BUILD) run --rm --no-deps storybook
 
-build-front-staging:
+build-front:
+ifeq ($(NODE_ENV),staging)
 	$(DOCKER_COMPOSE_BUILD) run --rm --no-deps front_staging
+else
+    ifeq ($(NODE_ENV),production)
+		$(DOCKER_COMPOSE_BUILD) run --rm --no-deps front_prod
+    else
+		$(DOCKER_COMPOSE_BUILD) run --rm --no-deps front_dev
+    endif
+endif
 
-build-admin-staging:
+build-admin:
+ifeq ($(NODE_ENV), staging)
 	$(DOCKER_COMPOSE_BUILD) run --rm --no-deps admin_staging
+else
+    ifeq ($(NODE_ENV), production)
+		$(DOCKER_COMPOSE_BUILD) run --rm --no-deps admin_prod
+    else
+		$(DOCKER_COMPOSE_BUILD) run --rm --no-deps admin_dev
+    endif
+endif
 
 build-api:
 	$(DOCKER_COMPOSE) run --rm --no-deps api npm run build
-
-build-front-dev:
-	$(DOCKER_COMPOSE_BUILD) run --rm --no-deps front_dev
-
-build-admin-dev:
-	$(DOCKER_COMPOSE_BUILD) run --rm --no-deps admin_dev
 
 build-dev: build-front-dev build-admin-dev
